Config Spring Security, Login Service complete

+++ BackEnd/ModelTradeAPI/src/main/java/nlu/modeltradeapi/services/implement/UserService.java
@@ -5,12 +5,15 @@ import nlu.modeltradeapi.dtos.requestdto.user.UserRegisterRequestDTO;
 import nlu.modeltradeapi.dtos.requestdto.user.UserUpdateRequestDTO;
 import nlu.modeltradeapi.entities.ActiveOTPUser;
 import nlu.modeltradeapi.entities.User;
+import nlu.modeltradeapi.exceptions.CustomException;
 import nlu.modeltradeapi.repository.ActiveOTPUserRepository;
 import nlu.modeltradeapi.repository.UserRepository;
 import nlu.modeltradeapi.services.template.IUserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSenderImpl;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -28,15 +31,17 @@ public class UserService implements IUserService {
     @Autowired
     private JavaMailSenderImpl mailSender;
 
+    private final PasswordEncoder passwordEncoder= new BCryptPasswordEncoder(10);
+
     @Override
     public User registerUser(UserRegisterRequestDTO registerRequest) {
         System.out.println(registerRequest.getUserName());
 
         // Kiểm tra username đã tồn tại
-        if (userRepository.findByUserName(registerRequest.getUserName()).isPresent()) {
+        if (userRepository.existsByUserName(registerRequest.getUserName())) {
             throw new RuntimeException("Username already exists");
         }
-        if (userRepository.findByEmail(registerRequest.getEmail()).isPresent()) {
+        if (userRepository.existsByEmail(registerRequest.getEmail())) {
             throw new RuntimeException("Email already exists");
         }
 
@@ -52,7 +57,6 @@ public class UserService implements IUserService {
                 .build();
         System.out.println(user.getUserName()+user.getEmail()+user.getPassword()+user.getPhoneNumber()+user.getDateOfBirth());
 
-        PasswordEncoder passwordEncoder= new BCryptPasswordEncoder(10);
         user.setPassword(passwordEncoder.encode(user.getPassword()));
         User savedUser = userRepository.save(user);
 
@@ -97,18 +101,29 @@ public class UserService implements IUserService {
     }
 
     @Override
-    public User updateUser(String userId, UserUpdateRequestDTO updateRequest) {
-        User user = getUserById(userId);
-        user.setName(updateRequest.getName());
-        user.setEmail(updateRequest.getEmail());
-        user.setPassword(updateRequest.getPassword());
-        user.setPhoneNumber(updateRequest.getPhoneNumber());
-        user.setDateOfBirth(updateRequest.getDateOfBirth());
+    public User updateUser(UserUpdateRequestDTO updateRequest) {
+        UserDetails userTrue = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        var userValue = userRepository.findByUserName(userTrue.getUsername()).orElseThrow(() -> new RuntimeException("User không tồn tại"));
+        User user = User.builder()
+                .userId(userValue.getUserId())
+                .userName(userValue.getUserName())
+                .email(updateRequest.getEmail())
+                .password(passwordEncoder.encode(updateRequest.getPassword()))
+                .name(updateRequest.getName())
+                .phoneNumber(updateRequest.getPhoneNumber())
+                .dateOfBirth(updateRequest.getDateOfBirth())
+                .createdDate(userValue.getCreatedDate())
+                .active(userValue.isActive())
+                .isDelete(updateRequest.isDeleted())
+                .build();
         return userRepository.save(user);
     }
 
     @Override
-    public void deleteUser(String userId) {
+    public void deleteUser(String userId) throws CustomException{
+        if(!userRepository.existsById(userId)) {
+            throw new CustomException("User not found");
+        }
         userRepository.deleteById(userId);
     }
